added an orders list to Mongo and fetching shipping rates from the ShipEngine api

+++ client/src/components/Layout.tsx
@@ -12,13 +12,14 @@ import HamburgerMenu from "./HamburgerMenu";
 import axios from "axios";
 
 export default function Layout() {
-    const { isAuthenticated, user } = useAuth0();
+    const { isAuthenticated, user, isLoading } = useAuth0();
     const [formData, setFormData] = useState<any>("");
     const [isCheckoutModalOpen, setIsCheckoutModalOpen] = useState(false);
     const [isHamburgerOpen, setIsHamburgerOpen] = useState<boolean>(false);
     const [cart, setCart] = useState<any>(
         JSON.parse(sessionStorage.getItem("cart") ?? "{}")
     );
+    const [isCartLoaded, setIsCartLoaded] = useState<boolean>(false);
     const [isCartOpen, setIsCartOpen] = useState<any>(false);
     const NavbarProps = {
         cart: cart,
@@ -28,15 +29,23 @@ export default function Layout() {
         isHamburgerOpen: isHamburgerOpen,
         setIsHamburgerOpen: setIsHamburgerOpen,
     };
+    const [shippingPrice, setShippingPrice] = useState<any>("");
+    const [isShippingPriceLoaded, setIsShippingPriceLoaded] =
+        useState<boolean>(false);
 
-    const postData = async () => {
+    const postOrder = async () => {
         const dataPosted = {
             name: formData.name,
             email: formData.email,
             phoneNumber: formData.phoneNumber,
             address: formData.address,
+            credit: formData?.credit,
+            cash: formData?.cash,
         };
-        await axios.post("http://localhost:4000/form", dataPosted);
+        await axios.post("http://localhost:4000/form", [
+            dataPosted,
+            user?.email,
+        ]);
     };
 
     const addUser = async () => {
@@ -53,24 +62,52 @@ export default function Layout() {
     };
 
     const getCart = async () => {
-        axios.get(`http://localhost:4000/cart/${user?.email}`)
-            .then(function (response) {
-                const data = response.data
-                const fixedData = data.map((obj:any) => {
-                    const productId = obj.id
+        axios
+            .get(`http://localhost:4000/cart/${user?.email}`)
+            .then((res) => {
+                const data = res.data;
+                const fixedData = data.map((obj: any) => {
+                    const productId = obj.id;
                     const { ...rest } = obj;
                     return { [productId]: rest };
                 });
-                const finalObj = {}
-                for(let i = 0; i < fixedData.length; i++ ) {
+                const finalObj = {};
+                for (let i = 0; i < fixedData.length; i++) {
                     Object.assign(finalObj, fixedData[i]);
                 }
                 setCart(finalObj);
+                setIsCartLoaded(true);
             })
-            .catch(function (error) {
+            .catch((error) => {
                 if (error) {
                     console.log(error);
                 }
+            });
+    };
+
+    const getShippingPrice = async () => {
+        console.log("getting rates");
+        console.log(formData.address);
+        axios
+            .get(
+                `http://localhost:4000/rates:${JSON.stringify(
+                    formData.address
+                )}`
+            )
+            .then((res) => {
+                console.log(res.data.shippingRate.rateResponse.rates);
+                const rates = res.data.shippingRate.rateResponse.rates;
+                const mappedRates = rates.map((data: any) => {
+                    return {
+                        shippingAmount: data.shippingAmount,
+                        deliveryDays: data.deliveryDays,
+                        estimatedDeliveryDate: data.estimatedDeliveryDate,
+                        serviceType: data.serviceType,
+                    };
+                });
+                console.log(mappedRates);
+                setShippingPrice(mappedRates);
+                setIsShippingPriceLoaded(true);
             });
     };
 
@@ -82,24 +119,46 @@ export default function Layout() {
     }, [isAuthenticated]);
 
     useEffect(() => {
-        if (!isAuthenticated) {
-            if (JSON.stringify(cart) !== sessionStorage.getItem("cart")) {
-                sessionStorage.setItem("cart", JSON.stringify(cart));
-            }
-        } else {
-            sessionStorage.clear();
-            updateCart();
+        if (formData) {
+            axios
+                .get(
+                    `http://localhost:4000/rates:${JSON.stringify(
+                        formData.address
+                    )}`
+                )
+                .then((res) => {
+                    console.log(res.data);
+                });
         }
-        console.log(cart)
-    }, [cart, isAuthenticated, user]);
+        // if(formData) {
+        //     getShippingPrice()
+        // }
+    }, [formData]);
 
     useEffect(() => {
-        updateCart()
-    }, [setIsCartOpen])
+        if (!isAuthenticated && !isLoading) {
+            if (JSON.stringify(cart) !== sessionStorage.getItem("cart")) {
+                sessionStorage.setItem("cart", JSON.stringify(cart));
+                setIsCartLoaded(true);
+            }
+        } else {
+            if (isCartLoaded) {
+                sessionStorage.clear();
+                updateCart();
+                setIsShippingPriceLoaded(false);
+            }
+        }
+    }, [cart, isAuthenticated, user, isLoading, isCartLoaded]);
+
+    // useEffect(() => {
+    //     isCartLoaded && updateCart()
+    // }, [isCartOpen])
 
     useEffect(() => {
         if (formData) {
-            postData();
+            postOrder();
+            getShippingPrice();
+            setFormData(undefined);
         }
     }, [formData]);
 
@@ -126,6 +185,7 @@ export default function Layout() {
                     setFormData,
                     isCheckoutModalOpen,
                     setIsCheckoutModalOpen,
+                    isCartLoaded,
                 }}
             />
             <Footer />
@@ -137,3 +197,4 @@ export default function Layout() {
 export function useLayoutOutletContext() {
     return useOutletContext<any>();
 }
+//
